+++ command/inspect.go
@@ -42,7 +42,12 @@ func (c *InspectCommand) Synopsis() string {
 }
 
 func (c *InspectCommand) AutocompleteFlags() complete.Flags {
-	return nil
+	return mergeAutocompleteFlags(c.Meta.AutocompleteFlags(FlagSetClient),
+		complete.Flags{
+			"-version": complete.PredictAnything,
+			"-json":    complete.PredictNothing,
+			"-t":       complete.PredictAnything,
+		})
 }
 
 func (c *InspectCommand) AutocompleteArgs() complete.Predictor {