+++ command/job_status.go
@@ -61,7 +61,13 @@ func (c *JobStatusCommand) Synopsis() string {
 }
 
 func (c *JobStatusCommand) AutocompleteFlags() complete.Flags {
-	return nil
+	return mergeAutocompleteFlags(c.Meta.AutocompleteFlags(FlagSetClient),
+		complete.Flags{
+			"-all-allocs": complete.PredictNothing,
+			"-evals":      complete.PredictNothing,
+			"-short":      complete.PredictNothing,
+			"-verbose":    complete.PredictNothing,
+		})
 }
 
 func (c *JobStatusCommand) AutocompleteArgs() complete.Predictor {