+++ command/node_drain.go
@@ -46,7 +46,13 @@ func (c *NodeDrainCommand) Synopsis() string {
 }
 
 func (c *NodeDrainCommand) AutocompleteFlags() complete.Flags {
-	return nil
+	return mergeAutocompleteFlags(c.Meta.AutocompleteFlags(FlagSetClient),
+		complete.Flags{
+			"-disable": complete.PredictNothing,
+			"-enable":  complete.PredictNothing,
+			"-self":    complete.PredictNothing,
+			"-yes":     complete.PredictNothing,
+		})
 }
 
 func (c *NodeDrainCommand) AutocompleteArgs() complete.Predictor {