+++ command/run.go
@@ -98,7 +98,14 @@ func (c *RunCommand) Synopsis() string {
 }
 
 func (c *RunCommand) AutocompleteFlags() complete.Flags {
-	return nil
+	return mergeAutocompleteFlags(c.Meta.AutocompleteFlags(FlagSetClient),
+		complete.Flags{
+			"-check-index": complete.PredictNothing,
+			"-detach":      complete.PredictNothing,
+			"-verbose":     complete.PredictNothing,
+			"-vault-token": complete.PredictAnything,
+			"-output":      complete.PredictNothing,
+		})
 }
 
 func (c *RunCommand) AutocompleteArgs() complete.Predictor {